Security check on parser and device
If an authenticated user doesn't own a device or a parser, he can parse all messages of the device related to a parser by knowing the parserId or parse all messages of the device by knowing the deviceId.
Check parser's own and device's own before parse messages. Only the owner must use the parser.